fix: inputField, tabs로 rename 및 tab 구현, 유연하게 전체적인 코드 수정
