+++ lib/presentation/component/input_field.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/material.dart';
+
+import 'package:recipe_app/ui/color_style.dart';
+import 'package:recipe_app/ui/text_font_style.dart';
+
+class InputField extends StatelessWidget {
+  final String label;
+  final String placeholder;
+  final void Function(String value) onValueChange;
+
+  const InputField({
+    super.key,
+    required this.label,
+    required this.placeholder,
+    required this.onValueChange,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 30),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(label, style: TextFontStyle.smallRegular()),
+          Container(
+            padding: EdgeInsets.only(top: 5),
+            width: double.infinity,
+            child: TextField(
+              style: TextFontStyle.smallRegular(color: ColorStyle.black),
+              decoration: InputDecoration(
+                contentPadding: EdgeInsets.all(20),
+                hintText: placeholder,
+                hintStyle: TextFontStyle.smallRegular(color: ColorStyle.gray4),
+                labelStyle: TextFontStyle.smallRegular(),
+                enabledBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(10),
+                  borderSide: BorderSide(color: ColorStyle.gray4, width: 1.5),
+                ),
+                focusedBorder: OutlineInputBorder(
+                  borderRadius: BorderRadius.circular(10),
+                  borderSide: BorderSide(
+                    color: ColorStyle.primary80,
+                    width: 1.5,
+                  ),
+                ),
+              ),
+              onChanged: (value) {
+                onValueChange(value);
+              },
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
