+++ pubspec.yaml
@@ -35,6 +35,9 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   google_fonts: ^6.2.1
+  freezed: ^3.0.6
+  freezed_annotation: ^3.0.0
+  build_runner: ^2.4.15
 
 dev_dependencies:
   flutter_test:
